requirements.txt
flask
Flask-Mail
pymongo
smtplib
PyPDF2
google-generativeai
python-dotenv
